back/requirements.txt
fastapi
uvicorn

tortoise-orm
asyncpg